Refactor: Move secret provider logic to shared/core
Relocate secret provider components, including `ISecretProvider`,
`FileSystemSecretProvider`, `VaultSecretProvider`, and the
`getSecretProviderInstance` utility, from `server/src/lib/secrets`
to the `shared/core` directory.

This change centralizes secret management, improving code
organization and enabling reusability. Import paths in dependent
server files have been updated to reflect the new location.
