+++ shared/core/ISecretProvider.ts
