+++ shared/core/index.ts
@@ -1,1 +1,5 @@
 export { getSecret } from './getSecret.js';
+export type { ISecretProvider } from './ISecretProvider.js';
+export { FileSystemSecretProvider } from './FileSystemSecretProvider.js';
+export { VaultSecretProvider } from './VaultSecretProvider.js';
+export { getSecretProviderInstance } from './secretProvider.js';
